refactor: move página de registros para dentro de /dashboard para roteamento intuitivo
