+++ src/views/pages/dashboard/registers/components/RegisterCard/index.tsx
