+++ src/views/pages/dashboard/registers/components/RegisterCard/styles.ts
