+++ src/views/pages/dashboard/registers/components/RegisterItem/index.tsx
