+++ src/views/pages/dashboard/registers/components/RegisterItem/styles.ts
