+++ src/views/pages/dashboard/registers/index.tsx
