+++ src/views/pages/dashboard/registers/styles.ts
